2007-06-06  Hans Boehm <Hans.Boehm@hp.com>
	* Makefile.am, Makefile.direct: Add NT_X64_STATIC_THREADS_MAKEFILE.
	* Makefile.in: Regenerate.
	* NT_X64_STATIC_THREADS_MAKEFILE: Fix warning flags.
	* allochblk.c, alloc.c, blacklst.c, dbg_malc.c, dyn_load.c,
	finalize.c, headers.c, mach_dep.c, malloc.c, mark.c, misc.c,
	obj_map.c, os_dep.c, ptr_chck.c, reclaim.c, typd_mlc.c,
	win32_threads.c, cord/de_win.c, include/gc_mark.h,
	include/private/gc_hdrs.h, include/private/gc_pmark.h,
	include/private/gc_priv.h, tests/test_cpp.cc:
	Replace old style function declarations.  Clean up integral types.
	Remove register declarations.  The change in malloc.c and the
	"int descr" declaration in mark.c are the most likely to have
	been real bugs outside of win64.
	* msvc_dbg.c: Disable on win64.
	* win32_threads.c: Add AMD64 support.
	* include/gc.h: no backtrace on AMD64 for now.

+++ cord/de_win.c
@@ -238,7 +238,7 @@ LRESULT CALLBACK WndProc (HWND hwnd, UINT message,
       	   if (wParam == QUIT) {
       	       SendMessage( hwnd, WM_CLOSE, 0, 0L );
       	   } else {
-      	       do_command(wParam);
+      	       do_command((int)wParam);
       	   }
       	   return(0);
       
@@ -323,13 +323,14 @@ LRESULT CALLBACK WndProc (HWND hwnd, UINT message,
       	           SetTextColor(dc, GetSysColor(COLOR_WINDOWTEXT));
       	           
       	           TextOut(dc, this_line.left, this_line.top,
-      	           	   plain, len);
-      	           TextOut(dc, this_line.left + len * char_width, this_line.top,
-      	           	   blanks, COLS - len);
+      	           	   plain, (int)len);
+      	           TextOut(dc, this_line.left + (int)len * char_width,
+		   	   this_line.top,
+      	           	   blanks, (int)(COLS - len));
       	           SetBkMode(dc, TRANSPARENT);
       	           SetTextColor(dc, RED);
       	           TextOut(dc, this_line.left, this_line.top,
-      	           	   control, strlen(control));
+      	           	   control, (int)strlen(control));
       	       }
       	   }
       	   EndPaint(hwnd, &ps);
